+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.0.0-rc.40",
-  genVersion: "2.755.6",
-  userAgent: "speakeasy-sdk/typescript 4.0.0-rc.40 2.755.6 2.0 @orq-ai/node",
+  sdkVersion: "4.0.0-rc.41",
+  genVersion: "2.755.9",
+  userAgent: "speakeasy-sdk/typescript 4.0.0-rc.41 2.755.9 2.0 @orq-ai/node",
 } as const;